seller screen working fine
